Add OM_OUTLINE_OVERLAP_COUNT parameter and remove outline_count limit
The mower has a tendency to miss a small patch when turning at the
end of the fill lines.
This allows the fill path to overlap the given number of outlines to
ensure this patch is located on an already mowed area.

Also remove the outline count limit to allow a large number of outlines
to be used to generate something similar to a concentric pattern.

Requires https://github.com/ClemensElflein/slic3r_coverage_planner/pull/2